Henry jk

+++ src/main/java/seedu/address/ui/MoreDetailsPanel.java
@@ -150,11 +150,11 @@ public class MoreDetailsPanel extends UiPart<Region> {
         label.setStyle(style);
         components.add(label, 3, row);
 
-        float assignmentWeight;
-        float assignmentMark;
-        float assignmentMaxMark;
-        float totalWeight = 0;
-        float weightedMarks = 0;
+        double assignmentWeight;
+        double assignmentMark;
+        double assignmentMaxMark;
+        double totalWeight = 0;
+        double weightedMarks = 0;
 
         Assignment assignment;
         for (int i = 0; i < assignmentList.size(); i++) {
